requirements.txt
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
Markdown==3.2.2
django-filter==2.4.0